refactor(init): inline starter handoff into the init command
Per AGENTS.md file organization rules:
- Move src/lib/starter.ts logic into src/commands/init.ts, its only
  caller. The lib layer must stay application-unaware; this code throws
  CommandError and prints CLI remediation messages.
- Remove hasPreviewByURL/removePreviewsByURL from the core client and
  inline the orchestration at the call site. Clients are network calls
  only.
- Remove test/starter-handoff.test.ts and test/repository-client.test.ts.
  Tests mirror src/commands/ one-to-one and never import from src/.
